+++ CHANGELOG.md
@@ -25,6 +25,12 @@ when it reaches v1, but until then consider all changes as possibly breaking.
 - Removed frontend alerts. From now on the concept of "alert" pertains only to
   messages from the backend.
 
+### Fixed
+
+- MIDI mapping a slider that only functioned as a modulation source did not
+  result in the destination parameter being modulated because the proxy is not
+  known as a dependency.
+
 ## [0.7.0] - 2025-04-10
 
 ### Added